+++ diesel/src/pg/expression/helper_types.rs
@@ -4,8 +4,7 @@ use crate::expression::grouped::Grouped;
 use crate::expression_methods::JsonIndex;
 use crate::pg::expression::expression_methods::private::JsonRemoveIndex;
 use crate::pg::types::sql_types::Array;
-use crate::sql_types::Nullable;
-use crate::sql_types::{Inet, Integer, VarChar};
+use crate::sql_types::{Inet, Integer, Nullable, VarChar};
 
 /// The return type of [`lhs.ilike(rhs)`](super::expression_methods::PgTextExpressionMethods::ilike)
 #[cfg(feature = "postgres_backend")]