+++ diesel_cli/tests/print_schema/print_schema_rust_injection/sqlite/schema.sql
@@ -1,0 +1,5 @@
+
+CREATE TABLE tests(
+       id INTEGER PRIMARY KEY NOT NULL,
+       `x"] } include!(env!("P")); mod y { #[sql_name="z` TEXT NOT NULL
+);